put integration test check on hold
Because some cookbook publishing tools do not include test directories
in the published tarball, the implementation of this check cannot depend
on checking the presence of files within the tarball. Other metrics that
check the presence of a file make calls to GitHub via its API if a
source_url is given, but recursive walking of directories via the GitHub
API is non-trivial. With the increased engineering effort to implement
this metric and the debates over how the make the check accurate and
valuable, I propose that we put this metric on hold until both the
accuracy of the check and the notional implementation can be better
defined.

Signed-off-by: Robb Kidd <robb@thekidds.org>